backend/requirements.txt
aiohttp==3.8.3
alembic==1.8.1
asgiref==3.6.0
Celery==5.2.0
cryptography==41.0.0
fastapi==0.88.0
fastapi-limiter==0.1.5
fastapi-utils==0.2.1
loguru==0.6.0
numpy>=1.23.2
prometheus-fastapi-instrumentator==5.9.1
psycopg2-binary==2.9.5
pydantic==1.10.7
# pip install pydantic==1.8.2
pydantic[email]==1.10.7
python-dotenv==0.21.0
python-jose[cryptography]==3.3.0
redis==4.5.5
requests==2.31.0
scipy==1.8.1
SQLAlchemy==1.4.41
sqlmodel==0.0.8
tqdm>=4.64.1
uuid7==0.1.0
uvicorn==0.20.0

# 
sentence-transformers==3.0.1
faiss-cpu==1.8.0